fix: price Long -> Double 타입으로 변경 대공사 작업

+++ backend/funco-trade/src/main/java/com/found_404/funco/feignClient/dto/FollowTradeRequest.java
@@ -13,7 +13,7 @@ public record FollowTradeRequest(
         TradeType tradeType,
         Double volume,
         Long orderCash,
-        Long price
+        Double price
 ) {
     public static FollowTradeRequest fromEntity(Trade trade) {
         return FollowTradeRequest.builder()
